Renamed file and code improvements
